refactored schemas
