+++ benchmarks/configs/experiment/document_classification_imdb_multi_label_v1.yaml
@@ -43,7 +43,7 @@ system_prompt_variables:
   project_name: Movie Genres
   project_description: An analysis of movie genres based on IMDb movie descriptions.
 
-schema: document_classification_schema.IMDBMultiLabelClassificationSchemaV1
+answer_schema: document_classification_schema.IMDBMultiLabelClassificationSchemaV1
 
 metrics:
   - multilabel_weighted_metrics
