+++ benchmarks/configs/experiment/document_classification_tagesschau_coarse_v1.yaml
@@ -22,7 +22,7 @@ system_prompt_variables:
   project_name: Tagesschau
   project_description: Eine Analyse der Themen in Deutschland basierend auf den Nachrichten und Artikeln der Tagesschau.
 
-schema: document_classification_schema.TagesschauCoarseClassificationSchemaV1
+answer_schema: document_classification_schema.TagesschauCoarseClassificationSchemaV1
 
 metrics:
   - classification_macro_metrics
