+++ benchmarks/configs/experiment/extractive_qa_germanquad_v1.yaml
@@ -9,7 +9,7 @@ prompt_template: extractive_qa_v1_de.j2
 
 system_prompt_template: extractive_qa_system_de.j2
 
-schema: extractive_qa_schema.ExtractiveQAAnswerSchemaV1
+answer_schema: extractive_qa_schema.ExtractiveQAAnswerSchemaV1
 
 metrics:
   - extractive_qa_squad2_metrics
