+++ benchmarks/configs/experiment/extractive_qa_squad2_v1.yaml
@@ -11,7 +11,7 @@ prompt_variables:
 
 system_prompt_template: extractive_qa_system_en.j2
 
-schema: extractive_qa_schema.ExtractiveQAAnswerSchemaV1
+answer_schema: extractive_qa_schema.ExtractiveQAAnswerSchemaV1
 
 metrics:
   - extractive_qa_squad2_metrics
