+++ benchmarks/configs/experiment/sequential_sentence_classification_csabstruct_v1.yaml
@@ -29,7 +29,7 @@ system_prompt_variables:
     - >-
       result - Presents the findings or outcomes of the research. This often includes statistical data, tables, and figures.
 
-schema: sequential_sentence_classification_schema.CSABStructSequentialSentenceClassificationSchemaV1
+answer_schema: sequential_sentence_classification_schema.CSABStructSequentialSentenceClassificationSchemaV1
 
 metrics:
   - sequential_sentence_classification_metrics
