+++ benchmarks/configs/experiment/span_classification_fewnerd_coarse_v1.yaml
@@ -17,7 +17,7 @@ prompt_variables:
 
 system_prompt_template: span_classification_system_en.j2
 
-schema: span_classification_schema.FewnerdCoarseSpanClassificationSchemaV1
+answer_schema: span_classification_schema.FewnerdCoarseSpanClassificationSchemaV1
 
 metrics:
   - span_classification_metrics
