+++ benchmarks/src/evaluation/metric_base.py
@@ -2,8 +2,8 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Generic, Sequence, TypeVar, cast, get_args, get_origin
 
-from schemas.answer_schema import BaseAnswerSchema
-from schemas.reference_schema import BaseReferenceSchema
+from schemas.prediction.prediction_schema import BaseAnswerSchema
+from schemas.reference.reference_schema import BaseReferenceSchema
 
 logger = logging.getLogger(__name__)
