+++ benchmarks/src/schemas/config/README.md
@@ -1,0 +1,3 @@
+# Config Schemas
+
+This diretory contains schemas for all files located in /benchmarks/configs. The schemas validate the Hydra configs.
