+++ benchmarks/src/schemas/prediction/document_classification_schema.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from pydantic import Field
 
-from schemas.answer_schema import (
+from schemas.prediction.prediction_schema import (
     MultiLabelClassificationSchema,
     SingleLabelClassificationSchema,
 )
